Moved ITs to sub-package 'inttest'
