+++ src/test/resources-its/de/r3s6/maven/constcreator/inttest/FailureHandlingIT/emptyPropertyKey/src/main/java/de/tester/main/Tester.java
