+++ src/test/resources-its/de/r3s6/maven/constcreator/inttest/FailureHandlingIT/templateSyntaxError/settings.xml
