+++ src/test/resources-its/de/r3s6/maven/constcreator/inttest/ProjectBuildIT/testProject/src/main/resources/messages/messages.properties
