+++ src/test/resources-its/de/r3s6/maven/constcreator/inttest/ProjectBuildIT/testProject/src/test/java/de/tester/constants/values/TestCaseValuesTest.java
